order api and email template

+++ app/Http/Controllers/API/Frontend/WishlistController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\API\Frontend;
 use Illuminate\Http\Request;
 use App\Http\Controllers\API\BaseController as BaseController;
 use Auth;
+use Validator;
+use App\Product;
 class WishlistController extends BaseController
 {
     public function index(){
@@ -16,7 +18,13 @@ class WishlistController extends BaseController
     }
 
     public function store(Request $request){
-
+       if(empty($request->id)){
+         return $this->sendError(null,'Please provide data.');
+       }
+       $product = Product::find($request->id);
+       if(is_null($product)){
+         return $this->sendError(null,'Product not found.');
+       }
       $wishlists = Auth::User()->userwishlist()->attach($request->id);
 
       return $this->sendResponse($wishlists,'Product added in wishlist');
@@ -25,7 +33,9 @@ class WishlistController extends BaseController
 
 
     public function remove(Request $request){
-
+      if(empty($request->id)){
+        return $this->sendError(null,'Please provide data.');
+      }
      $wishlists = Auth::User()->userwishlist()->detach($request->id);
      if($wishlists==false){
        return $this->sendError(null,'Product not in wishlist');
